Move o componente produtos para uma pasta propria e atualiza os imports
