+++ src/App.jsx
@@ -3,7 +3,7 @@ import Header from "./components/Header/Header";
 import Carrinho from "./components/Carrinho/Carrinho";
 import Footer from "./components/Footer";
 import Carrossel from "./components/Carrossel/Carrossel";
-import Produtos from "./components/Produtos";
+import Produtos from "./components/Produtos/Produtos";
 import NavIcon from "./components/NavIcon";
 
 export default function App() {
